Address feedback: rename and simplify
- Rename the library to libatomic-fallback to be clear it is not a real atomic library.
- Replace real atomic implementation in atomic.cpp with a static assert that prevents building this library for targets that do support atomic operations.
